fix(migration):prevent tag failure
Problem: There can be multiple tags with the same name. Some of this is
because we rename the incoming repo to remove spaces. Regardless, the import
will fail if we try to create it more than once.

Solution: Make sure the tag doesn't already exist before attempting to create
another

Testing: